backend/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-ckeditor==6.0.0
django-cors-headers==3.6.0
django-js-asset==1.2.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==2.0.0
pytz==2020.5
sqlparse==0.4.1
